.venv-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiosqlite
fastapi-cors
python-decouple